+++ ComputingProject/ComputingProject/key.h
@@ -8,7 +8,7 @@
 
 class Key {
 public:
-	void createKey(float dx, float dy, Bitmap* newSprite) {
+	void createKey(int dx, int dy, Bitmap* newSprite) {
 		position.setPoint(dx, dy);
 		sprite = newSprite;
 		spriteWidth = sprite->getWidth();
@@ -30,7 +30,7 @@ private:
 	Vector2 position;
 	BoundingBox bBox;
 	Bitmap* sprite;
-	float spriteWidth, spriteHeight;
+	int spriteWidth, spriteHeight;
 };
 
 #endif
